project/requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
Requests==2.32.3
mysql-connector-python==8.1.0
